refactor and some code
