--- libs/rules/prefer-react-fc/index.ts
@@ -1,1 +1,0 @@
-export * from "./prefer-react-fc";
