--- libs/rules/prefer-react-fc/prefer-react-fc.ts
@@ -1,12 +1,0 @@
-import { Rule } from "eslint";
-
-export const prefereReactFC: Rule.RuleModule = {
-  create: (context) => ({
-    "ImportDeclaration": (node) => {
-      context.report({
-        node,
-        message: 'Are you sure about this?'
-      })
-    }
-  })
-};
